faust_app/requirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiosignal==1.3.1
aiosmtplib==2.0.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
charset-normalizer==3.3.2
click==7.1.2
colorclass==2.2.2
colorlog==6.8.2
confluent-kafka==2.3.0
croniter==2.0.1
fastapi==0.103.0
faust==1.10.4
frozenlist==1.3.3
idna==3.6
importlib-metadata==6.7.0
kafka-python==1.4.7
loguru==0.7.2
mode==4.3.2
multidict==6.0.4
mypy-extensions==1.0.0
opentracing==1.3.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.4
robinhood-aiokafka==1.1.6
six==1.16.0
terminaltables==3.1.10
typing_extensions==4.7.1
venusian==1.2.0
yarl==1.9.4
zipp==3.15.0
